feat: clustering and object recognition
Tuned parameters to fit in the project setup.
1. Statistical Outlier Filtering:
Increase mean_k to 200 points to reject noise.

2. Passthrogh Filter:
Narrowed down table height range to [0.6, 0.75] in order to reject
outliers at left side.

3. RANSAC Plane Segmentation
Increased distance threshold to 0.02 so that the table can absorb some
noises of objects.

4. Euclidean Clustering
Decreased cluster tolerance to 0.1 because the objects are quite close
to each other. Because the objects are smaller, I also lower the
cluster size to [100, 500].